+++ nginx-auth/lua/config.lua
@@ -1,9 +1,6 @@
 -- -*- coding: utf-8 -*-
 -- (c) 2014 Andreas Motl, Elmyra UG <andreas.motl@elmyra.de>
 
--- our static user database ;]
-local users = { foo = "bar" };
-
 config = {
 
     auth = {
@@ -31,7 +28,6 @@ config = {
         cookie_name = 'Auth',
     },
 
-    users = users,
 };
 
 return config;