+++ nginx-auth/lua/lib/isis.lua
@@ -10,18 +10,25 @@ https://github.com/phaer/nginx-lua-auth
 
 ]]
 
+local cjson = require('cjson')
 local config = require('config')
-local users = config.users
+local util = require('lib/util')
 
 function authenticate_user(mode, username, password)
 
-    if users[username] ~= password then
+    ngx.req.set_header('Content-Type', 'application/json')
+    local response = ngx.location.capture(
+        '/api/identity/auth',
+        { method = ngx.HTTP_POST, body = cjson.encode({username=username, password=password}) }
+    )
+
+    if response.status == 200 then
+        log_auth_outcome(mode, true, username)
+        return cjson.decode(response.body)
+    else
         log_auth_outcome(mode, false, username)
-        return
     end
 
-    log_auth_outcome(mode, true, username)
-    return username
 end
 
 function log_auth_outcome(mode, success, username)
@@ -53,15 +60,52 @@ function get_basic_credentials()
 end
 
 -- TODO: encode_token(), decode_token()
--- TODO: put username or userid into token
+-- TODO: put username and/or userid into token
 
-function set_cookie()
+--[[
+TODO: maybe use Jwt for token encoding, see also:
+- https://github.com/Olivine-Labs/lua-jwt
+- https://github.com/webscriptio/lib/blob/master/jwt.lua
+]]
+
+function encode_token(data, timestamp)
+    local data = util.deepcopy(data)
+    local payload = cjson.encode(data)
+    local hmac = ngx.hmac_sha1(config.auth.hmac_secret, payload)
+
+    local payload_wire = ngx.encode_base64(payload)
+    local signature = ngx.encode_base64(hmac)
+    local token = timestamp .. ":" .. payload_wire .. ":" .. signature
+    return token
+end
+
+function decode_token(token)
+    -- Split token into timestamp, payload and HMAC signature
+
+    local timestamp, payload_wire, signature = token:match('(.+):(.+):(.+)')
+    timestamp = tonumber(timestamp)
+
+    local payload = ngx.decode_base64(payload_wire)
+    local hmac = ngx.decode_base64(signature)
+
+    local data = cjson.decode(payload)
+
+    local response = {
+        meta = {
+            hmac = hmac,
+            timestamp = timestamp,
+        },
+        data = data
+    }
+    return response
+end
+
+function set_cookie(user)
 
     local expires_after = config.auth.cookie_expiration
-
     local expiration = ngx.time() + expires_after
-    local signature = ngx.encode_base64(ngx.hmac_sha1(config.auth.hmac_secret, tostring(expiration)))
-    local token = expiration .. ":" .. signature
+    local token = encode_token(user, expiration)
+
     local cookie = config.auth.cookie_name .. "=" .. token .. "; "
     -- @TODO: Don't include subdomains
     cookie = cookie .. "Path=/; "
@@ -75,18 +119,18 @@ end
 function verify_cookie()
 
     local var_name = "cookie_" .. config.auth.cookie_name
-    local cookie = ngx.var[var_name]
+    local token = ngx.var[var_name]
 
     -- Check that the cookie exists.
-    if cookie ~= nil and cookie:find(":") ~= nil then
-        -- If there's a cookie, split off the HMAC signature
-        -- and timestamp.
-        local divider = cookie:find(":")
-        local hmac = ngx.decode_base64(cookie:sub(divider+1))
-        local timestamp = cookie:sub(0, divider-1)
+    if token ~= nil and token:find(":") ~= nil then
+
+        -- If there's a cookie, decode it.
+        local payload = decode_token(token)
+        local timestamp = payload.meta.timestamp
+        local user = payload.data
 
         -- Verify that the signature is valid and the token did not expire.
-        local ttl = tonumber(timestamp) - ngx.time()
+        local ttl = timestamp - ngx.time()
 
         -- TODO: prevent timing attacks, see also:
         --[[
@@ -97,9 +141,13 @@ function verify_cookie()
         http://www.levigross.com/2014/02/07/constant-time-comparison-functions-in-python-haskell-clojure-java-etc/
         see also: elmyra.web.identity
         ]]
-        if ngx.hmac_sha1(config.auth.hmac_secret, timestamp) == hmac and ttl >= 0 then
 
-            -- TODO: propagate userid/username to upstream service using http headers
+        if ttl >= 0 and encode_token(user, timestamp) == token then
+
+            -- propagate userid/username to upstream service via http headers
+            ngx.header['X-User-Id'] = user.userid
+            ngx.header['X-User-Username'] = user.username
+            ngx.header['X-User-Fullname'] = user.fullname
 
             -- automatic token renewal
             if ttl <= config.auth.cookie_renewal then