+++ elmyra.ip.access.epo/CHANGES.rst
@@ -13,6 +13,7 @@ development
 - ui/auth: enhance login box: integrate actions (login failed, register account) via email; rumble on error
 - auth: automatic cookie renewal, proper error responses for /api and /static routes
 - middleware: simple user identity subsystem
+- auth: authenticate against identity service, propagate userid/username to upstream service via http headers
 
 0.29.0
 ======